refactor(ibc-test): Move ckb to a standalone module
