+++ tools/ibc-test/src/tests/ckb/mod.rs
@@ -1,0 +1,2 @@
+/// Contains all ckb specific tests
+pub mod packet;
