+++ tools/ibc-test/src/tests/ibc/mod.rs
@@ -1,0 +1,2 @@
+/// Contains all IBC tests
+pub mod channel;
